# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: EDU-MENTOR API
  version: 0.1.0
  description: |
    Contrato inicial del piloto. La autorización combina roles, mentor assignments
    y ownership. Los módulos futuros están deliberadamente excluidos.
  x-contract-status: technical-baseline
servers:
  - url: /api/v1
    description: Same-origin reverse proxy
...
security:
  - accessCookie: []
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Listar usuarios
      operationId: listAdminUsers
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Usuarios de la organización.
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUser"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin]
      summary: Crear un usuario
      operationId: createAdminUser
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Usuario ficticio/real autorizado creado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/users/{userId}:
    patch:
      tags: [Admin]
      summary: Actualizar un usuario
      operationId: updateAdminUser
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Usuario actualizado dentro de la organización.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /admin/users/{userId}/password-reset:
    post:
      tags: [Admin]
      summary: Resetear contraseña de usuario
      operationId: resetUserPassword
      security:
        - accessCookie: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [temporaryPassword]
              properties:
                temporaryPassword:
                  type: string
                  minLength: 12
                  maxLength: 128
                  writeOnly: true
      responses:
        "204":
          description: Password temporal aplicado; sesiones revocadas.
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    accessCookie:
      type: apiKey
      in: cookie
      name: edu_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: edu_refresh
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Forbidden:
      description: Identidad válida sin permiso para la acción.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Recurso ausente o deliberadamente oculto por ownership.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Conflict:
      description: Conflicto de horario, versión, estado o idempotencia.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    Role:
      type: string
      enum: [PARTICIPANT, MENTOR, ADMIN]
...
    AdminUser:
      type: object
      additionalProperties: false
      required: [id, email, fullName, roles, isActive, mustChangePassword, version]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        fullName:
          type: string
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Role"
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
        version:
          type: integer
          minimum: 1
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [email, fullName, roles, temporaryPassword]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        fullName:
          type: string
          maxLength: 160
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Role"
        temporaryPassword:
          type: string
          minLength: 12
          maxLength: 128
          writeOnly: true
    UpdateUserRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      anyOf:
        - required: [fullName]
        - required: [roles]
        - required: [isActive]
      properties:
        fullName:
          type: string
          minLength: 1
          maxLength: 160
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Role"
        isActive:
          type: boolean
        expectedVersion:
          type: integer
          minimum: 1
...
    PageMeta:
      type: object
      additionalProperties: false
      required: [page, limit, total, hasNextPage]
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        hasNextPage:
          type: boolean
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, traceId]
          properties:
            code:
              type: string
              pattern: "^[A-Z][A-Z0-9_]+$"
            message:
              type: string
            traceId:
              type: string
            details:
              type: object
              additionalProperties: true
...